ai_person/requirements.txt
flask==2.3.2
qianfan-sdk==0.1.3
python-dotenv==1.0.0